chore: markdown is rendered in a separated component
